Deprecate USBPollerThread.
USBPollerThread expects an API which exists in only two standard python
object types: poll and epoll (the latter available on linux only).
But when using poll, the same long-pause issue as documented in USBPoller
happens, this time forbidding use of synchronous API.
Moreover, the presence of this class misleads users which would be better
served just calling handleEvents in a loop (either their main loop or a
separate thread).