Correct scaling of max power for Super Speed
The 'MaxPower' field of the Configuration descriptor is in units of 2mA
for everything pre-dating Super Speed and in units of 8mA for Super
Speed. Also, it is actually a current (measured in Amps), not a power
(measured in Watts); so the previous docstring was wrong/misleading when
it talked about mW (milliwatts).