runTestLibusb.sh: Allow specifying the python interpreter to test with.
travis-ci.org is now dead and I do not feel comfortable with the
permissions requested by travis-ci.com, so I guess I will be running the
tests locally. So I might as well make it easier to pick python versions.
So use virtualenv to get a fresh python environment to run tests in, and
run the tests (with only the versions I have available which do work)
before every release. Sadly, pypy does not make the cut for some boring
wheel installation reason.