usb1: Replace lambda in finalizer argument with partial
While it seems the lambda instance does not keep a reference to
setPollFDNotifiers's stack cell (which would cause the finalizer to keep a
reference to the object it is monitoring, preventing proper finalization)
in cPython 3.11, I am not confident this would be also the case in other
implementations and versions. I do not expect functools.partial instances
to keep any reference to their call stack, so they seem safer in such
reference-sensitive context.