usb1._libusb1: Rework __getLibrary
First prepare the list of filenames/filepaths to try loading, including
a possible fallback to find_library, and then try loading them.
This allows improving the error being raised so it can be better
representative of the attempted loads.
It also makes the code easier to read.
Only minor functional changes are intended: the exact exception type being
raised when find_library isn't used, and the content of the exception
raised otherwise.